refactor: keep system settings orchestration in core
